refactor: move client types next to Client class
