+++ src/types.ts
@@ -1,5 +1,4 @@
 export * from './types/base';
-export * from './types/client';
 export * from './types/request';
 export * from './types/response';
 export * from './types/query';
